Add instance for `Header`
It's unfortunate that we have to add this instance here,
but there are like three projects at play.

It doesn't make sense to push this into `servant`,
it almost surely wouldn't be accepted.

It doesn't make sense to push this into `servant-elm`.
That would be the ideal place for it,
but it would still be an orphan there.
If we're going to end up with an orhpan, why bother pushing it there?